move mapping to single mod
